refactor(desktop): split large frontend infra modules
