+++ apps/desktop/src/hooks/access/tauri/use-updater.ts
@@ -16,190 +16,28 @@ import {
   captureTelemetryException,
 } from "@/lib/integrations/telemetry/client";
 import { classifyTelemetryFailure } from "@/lib/domain/telemetry/failures";
+import {
+  clearDevUpdaterMockDownload,
+  DEV_UPDATER_MOCK_EVENT,
+  isDevUpdaterMockSupported,
+  readDevUpdaterMock,
+  startDevUpdaterMockDownload,
+  updateDevUpdaterMock,
+  writeDevUpdaterMock,
+  type DevUpdaterMockState,
+} from "./updater-dev-mock";
 
 const INITIAL_CHECK_DELAY_MS = 10_000;
 const CHECK_INTERVAL_MS = 21_600_000; // 6 hours
 const UPDATER_METADATA_KEY = "updater_metadata";
 const LEGACY_LAST_CHECKED_KEY = "updater_lastCheckedAt";
-const DEV_UPDATER_MOCK_KEY = "proliferate.dev.updaterMock";
-const DEV_UPDATER_MOCK_EVENT = "proliferate:dev-updater-mock";
-const DEV_UPDATER_MOCK_VERSION = "0.1.3";
-const DEV_UPDATER_MOCK_DOWNLOAD_DELAYS_MS = [200, 450, 700];
-const DEV_UPDATER_MOCK_DOWNLOAD_PROGRESS = [32, 68, 100];
 
 let checkInFlight = false;
 let autoCheckConsumerCount = 0;
 let stopAutoCheckScheduler: (() => void) | null = null;
-let mockDownloadTimeouts: number[] = [];
 
 interface UpdaterMetadata {
   lastCheckedAt: string | null;
-}
-
-type DevUpdaterMockPhase = Extract<UpdaterPhase, "available" | "downloading" | "ready">;
-
-interface DevUpdaterMockState {
-  phase: DevUpdaterMockPhase;
-  version: string;
-  downloadProgress: number | null;
-  restartPromptOpen: boolean;
-  lastCheckedAt: string | null;
-  errorMessage: string | null;
-}
-
-function isDevUpdaterMockSupported(): boolean {
-  return import.meta.env.DEV && typeof window !== "undefined";
-}
-
-function isDevUpdaterMockPhase(value: unknown): value is DevUpdaterMockPhase {
-  return value === "available" || value === "downloading" || value === "ready";
-}
-
-function normalizeDevUpdaterMock(raw: unknown): DevUpdaterMockState | null {
-  if (typeof raw === "string" && isDevUpdaterMockPhase(raw)) {
-    return {
-      phase: raw,
-      version: DEV_UPDATER_MOCK_VERSION,
-      downloadProgress: raw === "downloading" ? 0 : null,
-      restartPromptOpen: raw === "ready",
-      lastCheckedAt: null,
-      errorMessage: null,
-    };
-  }
-
-  if (!raw || typeof raw !== "object") {
-    return null;
-  }
-
-  const candidate = raw as Partial<DevUpdaterMockState>;
-  if (!isDevUpdaterMockPhase(candidate.phase)) {
-    return null;
-  }
-
-  return {
-    phase: candidate.phase,
-    version: candidate.version?.trim() || DEV_UPDATER_MOCK_VERSION,
-    downloadProgress:
-      candidate.phase === "downloading"
-        ? Math.max(0, Math.min(100, candidate.downloadProgress ?? 0))
-        : null,
-    restartPromptOpen:
-      typeof candidate.restartPromptOpen === "boolean"
-        ? candidate.restartPromptOpen
-        : candidate.phase === "ready",
-    lastCheckedAt:
-      typeof candidate.lastCheckedAt === "string" ? candidate.lastCheckedAt : null,
-    errorMessage:
-      typeof candidate.errorMessage === "string" ? candidate.errorMessage : null,
-  };
-}
-
-function readDevUpdaterMock(): DevUpdaterMockState | null {
-  if (!isDevUpdaterMockSupported()) {
-    return null;
-  }
-
-  const raw = window.localStorage.getItem(DEV_UPDATER_MOCK_KEY);
-  if (!raw) {
-    return null;
-  }
-
-  try {
-    return normalizeDevUpdaterMock(JSON.parse(raw));
-  } catch {
-    return normalizeDevUpdaterMock(raw);
-  }
-}
-
-function emitDevUpdaterMockChange(): void {
-  if (!isDevUpdaterMockSupported()) {
-    return;
-  }
-
-  window.dispatchEvent(new Event(DEV_UPDATER_MOCK_EVENT));
-}
-
-function clearDevUpdaterMockDownload(): void {
-  if (!isDevUpdaterMockSupported()) {
-    return;
-  }
-
-  for (const timeoutId of mockDownloadTimeouts) {
-    window.clearTimeout(timeoutId);
-  }
-  mockDownloadTimeouts = [];
-}
-
-function writeDevUpdaterMock(nextState: DevUpdaterMockState | null): void {
-  if (!isDevUpdaterMockSupported()) {
-    return;
-  }
-
-  if (!nextState) {
-    window.localStorage.removeItem(DEV_UPDATER_MOCK_KEY);
-    emitDevUpdaterMockChange();
-    return;
-  }
-
-  window.localStorage.setItem(DEV_UPDATER_MOCK_KEY, JSON.stringify(nextState));
-  emitDevUpdaterMockChange();
-}
-
-function updateDevUpdaterMock(
-  updater: (current: DevUpdaterMockState | null) => DevUpdaterMockState | null,
-): void {
-  writeDevUpdaterMock(updater(readDevUpdaterMock()));
-}
-
-function startDevUpdaterMockDownload(): void {
-  if (!isDevUpdaterMockSupported()) {
-    return;
-  }
-
-  const current = readDevUpdaterMock();
-  if (!current) {
-    return;
-  }
-
-  clearDevUpdaterMockDownload();
-  writeDevUpdaterMock({
-    ...current,
-    phase: "downloading",
-    downloadProgress: 0,
-    restartPromptOpen: false,
-    errorMessage: null,
-  });
-
-  for (let index = 0; index < DEV_UPDATER_MOCK_DOWNLOAD_PROGRESS.length; index += 1) {
-    const timeoutId = window.setTimeout(() => {
-      const latest = readDevUpdaterMock();
-      if (!latest) {
-        return;
-      }
-
-      if (index === DEV_UPDATER_MOCK_DOWNLOAD_PROGRESS.length - 1) {
-        writeDevUpdaterMock({
-          ...latest,
-          phase: "ready",
-          downloadProgress: null,
-          restartPromptOpen: true,
-          errorMessage: null,
-        });
-        clearDevUpdaterMockDownload();
-        return;
-      }
-
-      writeDevUpdaterMock({
-        ...latest,
-        phase: "downloading",
-        downloadProgress: DEV_UPDATER_MOCK_DOWNLOAD_PROGRESS[index],
-        restartPromptOpen: false,
-        errorMessage: null,
-      });
-    }, DEV_UPDATER_MOCK_DOWNLOAD_DELAYS_MS[index]);
-
-    mockDownloadTimeouts.push(timeoutId);
-  }
 }
 
 async function persistUpdaterMetadata(metadata: UpdaterMetadata): Promise<void> {
