+++ apps/desktop/src/lib/infra/terminals/terminal-close-intent.ts
@@ -1,0 +1,17 @@
+const intentionallyClosingTerminalIds = new Set<string>();
+
+export function markTerminalIntentionalClose(terminalId: string): void {
+  intentionallyClosingTerminalIds.add(terminalId);
+}
+
+export function clearTerminalIntentionalClose(terminalId: string): void {
+  intentionallyClosingTerminalIds.delete(terminalId);
+}
+
+export function isTerminalIntentionalClose(terminalId: string): boolean {
+  return intentionallyClosingTerminalIds.has(terminalId);
+}
+
+export function resetTerminalCloseIntentForTests(): void {
+  intentionallyClosingTerminalIds.clear();
+}
