+++ scripts/max_lines_allowlist.txt
@@ -29,10 +29,7 @@ apps/desktop/src/hooks/workspaces/cache/tabs/use-workspace-header-subagent-hiera
 apps/desktop/src/hooks/workspaces/workflows/tabs/use-workspace-shell-activation.ts 671 existing desktop hook oversized file
 apps/desktop/src/hooks/workspaces/workflows/use-workspace-bootstrap-actions.ts 899 existing desktop hook oversized file
 apps/desktop/src/lib/domain/workspaces/cloud/logical-workspaces.test.ts 759 existing desktop oversized test file
-apps/desktop/src/lib/infra/measurement/boot-stall-diagnostics.ts 1284 existing desktop oversized file
-apps/desktop/src/lib/integrations/auth/orchestration.ts 628 existing desktop oversized file
 apps/desktop/src/lib/workflows/mcp/connector-persistence.test.ts 704 existing desktop oversized test file
-apps/desktop/src/lib/workflows/mcp/connector-persistence.ts 638 existing desktop oversized file
 server/proliferate/auth/identity/service.py 809 existing auth service exceeds repo max-lines threshold
 server/proliferate/db/store/automations.py 801 server structure migration debt for oversized store file
 server/proliferate/db/store/billing.py 2602 server structure migration debt for oversized store file
